refactor(graph): Move graph logic to lib directory
- Refactor the project structure by moving graph-related logic from
  the `contexts` and `component` directories to a dedicated `lib`
  directory. This is intended to enhance the organization and clarity
  of the codebase.

- Update import statements across relevant files to use the new
  locations for the graph functions and types.

No breaking changes are introduced in this refactor.
